third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/onerror-event-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/access-control-basic-denied.cgi' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
This test that the error event is fired for XMLHttpRequests

PASS: error event fired.
